Refactor presences script into separate classes + extract some utilities into other files for reuse
